+++ Omega/src/com/saggitt/omega/gestures/gestures/LongPressGesture.kt
@@ -25,13 +25,10 @@ class LongPressGesture(controller: GestureController) : Gesture(controller) {
 
     private val handler by controller.createHandlerPref("pref_gesture_long_press",
             OpenOverviewGestureHandler(controller.launcher, null))
-
-    override fun getIsEnabled(): Boolean {
-        return true;
-    }
+    override val isEnabled = true
 
     override fun onEvent(): Boolean {
-        handler.onGestureTrigger(mController)
+        handler.onGestureTrigger(controller)
         return true
     }
 }